requirements.txt
flask==2.0.1
pandas==1.3.3
requests==2.26.0
kaggle==1.5.12
python-dotenv==0.19.0
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0 